refactor: rename into `select_builder`
